Tx_Rx
manifold subscriptions.....
limited testing, but passes all tests.
common_Rx channel is created at installation, this channel is intended for com bus(comunication bus) requests. in testing tab you provide a name, roles, a common_Tx( others common_Rx) and a channel type for the com bus.

most of subscriptions api is the same for Tx_Rx in this commit, but that will change.