Added a sidebar and drag/drop functionality
We can now drag cards from a side bar on the communities page to hover over a community. Dropping it will fire an event that is not yet configured to dispatch an action, because the action functionality doesn't exist yet in the manifold krl ruleset. The UI functionality is just about set up and ready to go to add the feature of putting things/communities into a community.